requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
mcp>=0.9.0
google-ads>=24.1.0
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0
